+++ applications/helpers/reverse_documents.py
