+++ applications/helpers/summaries.py
