+++ applications/views/documents.py
@@ -9,7 +9,7 @@ from django.views.generic import TemplateView
 from s3chunkuploader.file_handler import S3FileUploadHandler
 
 from applications.forms.end_user import attach_document_form, delete_document_confirmation_form
-from applications.libraries.reverse_documents import document_switch
+from applications.helpers.reverse_documents import document_switch
 from applications.services import add_document_data, download_document_from_s3
 from core.builtins.custom_tags import get_string
 from lite_forms.generators import form_page, error_page
