+++ applications/views/end_users.py
@@ -3,8 +3,8 @@ from django.urls import reverse_lazy
 from django.views.generic import TemplateView
 
 from applications.forms.end_user import new_end_user_forms
-from applications.libraries.check_your_answers_helpers import convert_end_user
-from applications.libraries.validate_status import check_all_parties_have_a_document
+from applications.helpers.check_your_answers import convert_end_user
+from applications.helpers.validate_status import check_all_parties_have_a_document
 from applications.services import (
     get_application,
     post_end_user,
